Update Core Extensions
