--- CoreComponent/CoreProject/coreProject.h
@@ -1,18 +1,0 @@
-//
-//  coreProject.h
-//  CoreProject
-//
-//  Created by Yannick Spreen on 5/8/21.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for CoreProject.
-FOUNDATION_EXPORT double coreProjectVersionNumber;
-
-//! Project version string for CoreProject.
-FOUNDATION_EXPORT const unsigned char coreProjectVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <coreProject/PublicHeader.h>
-
-
