+++ Sources/Services/ZLib.swift
@@ -22,21 +22,22 @@
 import Foundation
 import Compression
 
-func decompressString(_ data: Data) -> String {
-  return String(decoding: decompress(data), as: UTF8.self)
-}
+public class ZLib {
+    static func decompressString(_ data: Data) -> String {
+      return String(decoding: decompress(data), as: UTF8.self)
+    }
 
-public func decompress(_ data: Data) -> Data {
-  if data.count <= 2 {
-    return .init()
-  }
-  let size = 4 * data.count + 8 * 1024
-  let buffer = UnsafeMutablePointer<UInt8>.allocate(capacity: size)
-  let result = data.subdata(in: 2 ..< data.count).withUnsafeBytes {
-    let read = compression_decode_buffer(buffer, size, $0.baseAddress!.bindMemory(to: UInt8.self, capacity: 1),
-                                         data.count - 2, nil, COMPRESSION_ZLIB)
-    return Data(bytes: buffer, count: read)
-  } as Data
-  buffer.deallocate()
-  return result
+    public static func decompress(_ data: Data) -> Data {
+        if data.count <= 2 {
+          return .init()
+        }
+        let size = 4 * data.count + 8 * 1024
+        let buffer = UnsafeMutablePointer<UInt8>.allocate(capacity: size)
+        let result = data.subdata(in: 2 ..< data.count).withUnsafeBytes {
+          let read = compression_decode_buffer(buffer, size, $0.baseAddress!.bindMemory(to: UInt8.self, capacity: 1), data.count - 2, nil, COMPRESSION_ZLIB)
+          return Data(bytes: buffer, count: read)
+        } as Data
+        buffer.deallocate()
+        return result
+    }
 }
